move vite plugin
